// src/generated/models/Rpvms_contractsModel.ts
export const Rpvms_contractsrpvms_amended = {
  412900000: 'Yes',
  412900001: 'No',
  412900002: 'N_A',
  412900003: 'Unknown'
} as const;
export type Rpvms_contractsrpvms_amended = keyof typeof Rpvms_contractsrpvms_amended;
export const Rpvms_contractsrpvms_autorenew = {
  412900000: 'Yes',
  412900001: 'No',
  412900002: 'N_A',
  412900003: 'Unknown'
} as const;
export type Rpvms_contractsrpvms_autorenew = keyof typeof Rpvms_contractsrpvms_autorenew;
export const Rpvms_contractsrpvms_contractstatus = {
  412900000: 'Active',
  412900001: 'Expired',
  412900002: 'Terminated',
  412900003: 'Pending',
  412900004: 'UnderReview',
  412900005: 'Unknown'
} as const;
export type Rpvms_contractsrpvms_contractstatus = keyof typeof Rpvms_contractsrpvms_contractstatus;
export const Rpvms_contractsrpvms_contracttype = {
  412900000: 'MasterServicesAgreement',
  412900001: 'StatementofWork',
  412900002: 'OrderForm',
  412900003: 'LicenseAgreement',
  412900004: 'NDA',
  412900005: 'Amendment',
  412900006: 'BAA',
  412900007: 'Other'
} as const;
export type Rpvms_contractsrpvms_contracttype = keyof typeof Rpvms_contractsrpvms_contracttype;
export const Rpvms_contractsrpvms_terminationwithoutcause = {
  412900000: 'Yes',
  412900001: 'No',
  412900002: 'N_A',
  412900003: 'Unknown'
} as const;
export type Rpvms_contractsrpvms_terminationwithoutcause = keyof typeof Rpvms_contractsrpvms_terminationwithoutcause;