// src/generated/models/Rpvms_onetrustassessmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Rpvms_onetrustassessmentsrpvms_classification = {
  412900000: 'Clinical',
  412900001: 'ProfessionalServices',
  412900002: 'ITInfrastructure',
  412900003: 'Security',
  412900004: 'Telecom',
  412900005: 'Staffing',
  412900006: 'Other'
} as const;
export type Rpvms_onetrustassessmentsrpvms_classification = keyof typeof Rpvms_onetrustassessmentsrpvms_classification;
export const Rpvms_onetrustassessmentsrpvms_criticality = {
  412900000: '_1_Negligible',
  412900001: '_2_Low',
  412900002: '_3_Noticeable',
  412900003: '_4_Considerable',
  412900004: '_5_Catastrophic'
} as const;
export type Rpvms_onetrustassessmentsrpvms_criticality = keyof typeof Rpvms_onetrustassessmentsrpvms_criticality;
export const Rpvms_onetrustassessmentsrpvms_ephi = {
  412900000: 'Yes',
  412900001: 'No',
  412900002: 'N_A',
  412900003: 'Unknown'
} as const;
export type Rpvms_onetrustassessmentsrpvms_ephi = keyof typeof Rpvms_onetrustassessmentsrpvms_ephi;
export const Rpvms_onetrustassessmentsrpvms_systemaccess = {
  412900000: 'None',
  412900001: 'Read',
  412900002: 'Read_Write',
  412900003: 'Admin',
  412900004: 'Unknown'
} as const;
export type Rpvms_onetrustassessmentsrpvms_systemaccess = keyof typeof Rpvms_onetrustassessmentsrpvms_systemaccess;
export const Rpvms_onetrustassessmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Rpvms_onetrustassessmentsstatecode = keyof typeof Rpvms_onetrustassessmentsstatecode;
export const Rpvms_onetrustassessmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Rpvms_onetrustassessmentsstatuscode = keyof typeof Rpvms_onetrustassessmentsstatuscode;

export interface Rpvms_onetrustassessmentsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  rpvms_classification?: Rpvms_onetrustassessmentsrpvms_classification;
  rpvms_criticality?: Rpvms_onetrustassessmentsrpvms_criticality;
  rpvms_ephi?: Rpvms_onetrustassessmentsrpvms_ephi;
  rpvms_integrations?: string;
  rpvms_onetrustassessmentid: string;
  rpvms_onetrustassessmentname: string;
  rpvms_productservice?: string;
  rpvms_systemaccess?: Rpvms_onetrustassessmentsrpvms_systemaccess;
  rpvms_type?: string;
  "rpvms_VendorId@odata.bind": string;
  statecode: Rpvms_onetrustassessmentsstatecode;
  statuscode?: Rpvms_onetrustassessmentsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Rpvms_onetrustassessments extends Rpvms_onetrustassessmentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  rpvms_classificationname?: string;
  rpvms_criticalityname?: string;
  rpvms_ephiname?: string;
  rpvms_systemaccessname?: string;
  rpvms_vendoridname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  rpvms_vendorid?: object;
  _rpvms_vendorid_value?: string;
}