// src/generated/models/Rpvms_servicenowassessmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Rpvms_servicenowassessmentsrpvms_assessmenttype = {
  412900000: 'Criticality',
  412900001: 'Performance'
} as const;
export type Rpvms_servicenowassessmentsrpvms_assessmenttype = keyof typeof Rpvms_servicenowassessmentsrpvms_assessmenttype;
export const Rpvms_servicenowassessmentsrpvms_criticalitylevel = {
  412900000: '_1_Negligible',
  412900001: '_2_Low',
  412900002: '_3_Noticeable',
  412900003: '_4_Considerable',
  412900004: '_5_Catastrophic'
} as const;
export type Rpvms_servicenowassessmentsrpvms_criticalitylevel = keyof typeof Rpvms_servicenowassessmentsrpvms_criticalitylevel;
export const Rpvms_servicenowassessmentsrpvms_isbudgeted = {
  412900000: 'Yes',
  412900001: 'No',
  412900002: 'N_A',
  412900003: 'Unknown'
} as const;
export type Rpvms_servicenowassessmentsrpvms_isbudgeted = keyof typeof Rpvms_servicenowassessmentsrpvms_isbudgeted;
export const Rpvms_servicenowassessmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Rpvms_servicenowassessmentsstatecode = keyof typeof Rpvms_servicenowassessmentsstatecode;
export const Rpvms_servicenowassessmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Rpvms_servicenowassessmentsstatuscode = keyof typeof Rpvms_servicenowassessmentsstatuscode;

export interface Rpvms_servicenowassessmentsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  rpvms_assessmenttype: Rpvms_servicenowassessmentsrpvms_assessmenttype;
  rpvms_budgetedcostcenter?: string;
  rpvms_costcentercode?: string;
  rpvms_costcentername?: string;
  rpvms_criticalitylevel?: Rpvms_servicenowassessmentsrpvms_criticalitylevel;
  rpvms_impactedcostcenter?: string;
  rpvms_isbudgeted?: Rpvms_servicenowassessmentsrpvms_isbudgeted;
  rpvms_itsponsor?: string;
  rpvms_perfcontractchangereq?: string;
  rpvms_perfissuehandling?: string;
  rpvms_perfoverallreliability?: string;
  rpvms_perfsupportsatisfaction?: string;
  rpvms_productservicetype?: string;
  rpvms_relatedlocalpractice?: string;
  rpvms_relatedrpdepartment?: string;
  rpvms_requestscope?: string;
  rpvms_servicenowassessmentid: string;
  rpvms_servicenowassessmentname: string;
  rpvms_snnumber1?: string;
  "rpvms_VendorId@odata.bind": string;
  rpvms_vendormailingaddress?: string;
  rpvms_vendornameraw?: string;
  rpvms_vendorrepemail?: string;
  rpvms_vendorrepname?: string;
  rpvms_vendorreptitle?: string;
  statecode: Rpvms_servicenowassessmentsstatecode;
  statuscode?: Rpvms_servicenowassessmentsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Rpvms_servicenowassessments extends Rpvms_servicenowassessmentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  rpvms_assessmenttypename?: string;
  rpvms_criticalitylevelname?: string;
  rpvms_isbudgetedname?: string;
  rpvms_vendoridname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  rpvms_vendorid?: object;
  _rpvms_vendorid_value?: string;
}