// src/generated/models/Rpvms_suppliersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Rpvms_suppliersrpvms_isreseller = {
  0: 'No',
  1: 'Yes'
} as const;
export type Rpvms_suppliersrpvms_isreseller = keyof typeof Rpvms_suppliersrpvms_isreseller;
export const Rpvms_suppliersrpvms_tintype = {
  412900000: 'EIN',
  412900001: 'SSN',
  412900002: 'ITIN',
  412900003: 'Foreign',
  412900004: 'Unknown'
} as const;
export type Rpvms_suppliersrpvms_tintype = keyof typeof Rpvms_suppliersrpvms_tintype;
export const Rpvms_suppliersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Rpvms_suppliersstatecode = keyof typeof Rpvms_suppliersstatecode;
export const Rpvms_suppliersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Rpvms_suppliersstatuscode = keyof typeof Rpvms_suppliersstatuscode;

export interface Rpvms_suppliersBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  rpvms_isreseller?: Rpvms_suppliersrpvms_isreseller;
  rpvms_suppliercategory?: string;
  rpvms_supplierid: string;
  rpvms_suppliername: string;
  rpvms_taxid?: string;
  rpvms_tintype?: Rpvms_suppliersrpvms_tintype;
  statecode: Rpvms_suppliersstatecode;
  statuscode?: Rpvms_suppliersstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Rpvms_suppliers extends Rpvms_suppliersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  rpvms_isresellername?: string;
  rpvms_tintypename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}