// src/generated/models/Rpvms_vendorsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Rpvms_vendorsrpvms_activephiaccess = {
  412900000: 'Yes',
  412900001: 'No',
  412900002: 'N_A',
  412900003: 'Unknown'
} as const;
export type Rpvms_vendorsrpvms_activephiaccess = keyof typeof Rpvms_vendorsrpvms_activephiaccess;
export const Rpvms_vendorsrpvms_classification = {
  412900000: 'Clinical',
  412900001: 'ProfessionalServices',
  412900002: 'ITInfrastructure',
  412900003: 'Security',
  412900004: 'Telecom',
  412900005: 'Staffing',
  412900006: 'Other'
} as const;
export type Rpvms_vendorsrpvms_classification = keyof typeof Rpvms_vendorsrpvms_classification;
export const Rpvms_vendorsrpvms_commercialrole = {
  412900000: 'Vendor',
  412900001: 'Reseller',
  412900002: 'VAR',
  412900003: 'Distributor',
  412900004: 'Hybrid'
} as const;
export type Rpvms_vendorsrpvms_commercialrole = keyof typeof Rpvms_vendorsrpvms_commercialrole;
export const Rpvms_vendorsrpvms_isvar = {
  0: 'No',
  1: 'Yes'
} as const;
export type Rpvms_vendorsrpvms_isvar = keyof typeof Rpvms_vendorsrpvms_isvar;
export const Rpvms_vendorsrpvms_status = {
  412900000: 'Active',
  412900001: 'Inactive',
  412900002: 'Onboarding',
  412900003: 'Offboarded',
  412900004: 'UnderReview'
} as const;
export type Rpvms_vendorsrpvms_status = keyof typeof Rpvms_vendorsrpvms_status;
export const Rpvms_vendorsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Rpvms_vendorsstatecode = keyof typeof Rpvms_vendorsstatecode;
export const Rpvms_vendorsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Rpvms_vendorsstatuscode = keyof typeof Rpvms_vendorsstatuscode;

export interface Rpvms_vendorsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  rpvms_activephiaccess?: Rpvms_vendorsrpvms_activephiaccess;
  rpvms_categoryl1?: string;
  rpvms_categoryl2?: string;
  rpvms_classification?: Rpvms_vendorsrpvms_classification;
  rpvms_commercialrole?: Rpvms_vendorsrpvms_commercialrole;
  rpvms_isvar?: Rpvms_vendorsrpvms_isvar;
  rpvms_primaryoffering?: string;
  rpvms_status?: Rpvms_vendorsrpvms_status;
  rpvms_vendorid: string;
  rpvms_vendorname: string;
  statecode: Rpvms_vendorsstatecode;
  statuscode?: Rpvms_vendorsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Rpvms_vendors extends Rpvms_vendorsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  rpvms_activephiaccessname?: string;
  rpvms_classificationname?: string;
  rpvms_commercialrolename?: string;
  rpvms_isvarname?: string;
  rpvms_statusname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}